CHG: Reduce boiler plate code

+++ Topics/The for-loop/The count of numbers divisible by N/src/Main.java
@@ -3,13 +3,11 @@ import java.util.stream.IntStream;
 
 class Main {
     public static void main(String[] args) {
-        // put your code here
         Scanner scanner = new Scanner(System.in);
         int a = scanner.nextInt();
         int b = scanner.nextInt();
         int n = scanner.nextInt();
-        int count = (int) IntStream.rangeClosed(a, b).filter(i -> i % n == 0).count();
-
+        long count = IntStream.rangeClosed(a, b).filter(i -> i % n == 0).count();
         System.out.println(count);
     }
 }
